move full granularity test to another existing file
